requirements.txt
streamlit
langchain
langchain-community
langchain-chroma
langchain-openai
python-dotenv
tiktoken
openai